Move random test to its own file
