+++ src/tests/mod.rs
@@ -205,22 +205,7 @@ impl TestToken<'_> {
     }
 }
 
-#[test]
-fn test_random() {
-    let mut testtokn = TestToken::initialized("test_random.json", None);
-    let session = testtokn.get_session(false);
-
-    let data: &[u8] = &mut [0, 0, 0, 0];
-    let ret = fn_generate_random(
-        session,
-        data.as_ptr() as *mut u8,
-        data.len() as CK_ULONG,
-    );
-    assert_eq!(ret, CKR_OK);
-    assert_ne!(data, &[0, 0, 0, 0]);
-
-    testtokn.finalize();
-}
+mod random;
 
 fn test_login(name: &str) {
     let mut testtokn = TestToken::initialized(name, None);
